pyproject.toml
[project]
name = "todoist-mcp-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.4.0",
    "httpx>=0.25.0",
    "uvicorn>=0.23.0",
    "python-dotenv>=1.1.1",
]
